Fixes #634 Selecting language for voice input
