+++ app/src/common/shared/org/mozilla/vrbrowser/VRBrowserActivity.java
@@ -47,7 +47,7 @@ import org.mozilla.vrbrowser.ui.OffscreenDisplay;
 import org.mozilla.vrbrowser.ui.widgets.BookmarkListener;
 import org.mozilla.vrbrowser.ui.widgets.BookmarksWidget;
 import org.mozilla.vrbrowser.ui.widgets.BrowserWidget;
-import org.mozilla.vrbrowser.ui.widgets.CrashDialogWidget;
+import org.mozilla.vrbrowser.ui.widgets.dialogs.CrashDialogWidget;
 import org.mozilla.vrbrowser.ui.widgets.KeyboardWidget;
 import org.mozilla.vrbrowser.ui.widgets.NavigationBarWidget;
 import org.mozilla.vrbrowser.ui.widgets.RootWidget;
