+++ app/src/common/shared/org/mozilla/vrbrowser/ui/widgets/TrayWidget.java
@@ -19,6 +19,7 @@ import org.mozilla.vrbrowser.R;
 import org.mozilla.vrbrowser.audio.AudioEngine;
 import org.mozilla.vrbrowser.browser.SessionStore;
 import org.mozilla.vrbrowser.ui.views.UIButton;
+import org.mozilla.vrbrowser.ui.widgets.dialogs.SettingsWidget;
 
 import java.util.ArrayList;
 import java.util.Arrays;
