+++ app/src/main/res/drawable/ic_settings_display.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="200"
+    android:viewportHeight="200">
+    <path
+        android:fillColor="@color/fog"
+        android:pathData="M98.19,39.47C57,40.51 25.07,65.59 4.83,100 25.07,134.42 57,159.49 98.2,160.53c43.49,1.1 75.35,-25.1 97,-60.53C173.55,64.57 141.69,38.37 98.19,39.47ZM100,146a46,46 0,1 1,46 -46A46,46 0,0 1,100 146Z" />
+    <path
+        android:fillColor="@color/fog"
+        android:pathData="M99.12,100.75l12.13,-23.14A25.07,25.07 0,1 0,100 125.07h0A25.07,25.07 0,0 0,125 100a25.42,25.42 0,0 0,-0.68 -5.83Z" />
+</vector>
