+++ app/src/main/res/drawable/ic_settings_privacypolicy.xml
@@ -1,39 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-<item android:state_pressed="true">
-    <layer-list>
-        <item>
-            <vector android:height="24dp" android:viewportHeight="200"
-                android:viewportWidth="200" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-                <path android:fillColor="@color/fog" android:pathData="M69.28,82.82H79.86V66a20.14,20.14 0,1 1,40.28 0V82.82h10.58a27.1,27.1 0,0 1,6.45 0.78l2.7,0.66V66a39.87,39.87 0,0 0,-79.74 0V84.25l2.7,-0.66a27.61,27.61 0,0 1,6.45 -0.77Z"/>
-                <path android:fillColor="@color/fog" android:pathData="M109.51,133.28v14.43a9.52,9.52 0,0 1,-19 0L90.51,133.28a16.45,16.45 0,1 1,19 0ZM130.72,86.74L69.28,86.74a25,25 0,0 0,-25 25L44.28,146.4a27.51,27.51 0,0 0,27.51 27.51h56.44a27.51,27.51 0,0 0,27.51 -27.51L155.74,111.75A25,25 0,0 0,130.72 86.74Z"/>
-            </vector>
-        </item>
-
-    </layer-list>
-</item>
-<item android:state_hovered="true">
-    <layer-list>
-        <item>
-            <vector android:height="24dp" android:viewportHeight="200"
-                android:viewportWidth="200" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-                <path android:fillColor="@color/asphalt" android:pathData="M69.28,82.82H79.86V66a20.14,20.14 0,1 1,40.28 0V82.82h10.58a27.1,27.1 0,0 1,6.45 0.78l2.7,0.66V66a39.87,39.87 0,0 0,-79.74 0V84.25l2.7,-0.66a27.61,27.61 0,0 1,6.45 -0.77Z"/>
-                <path android:fillColor="@color/asphalt" android:pathData="M109.51,133.28v14.43a9.52,9.52 0,0 1,-19 0L90.51,133.28a16.45,16.45 0,1 1,19 0ZM130.72,86.74L69.28,86.74a25,25 0,0 0,-25 25L44.28,146.4a27.51,27.51 0,0 0,27.51 27.51h56.44a27.51,27.51 0,0 0,27.51 -27.51L155.74,111.75A25,25 0,0 0,130.72 86.74Z"/>
-            </vector>
-        </item>
-
-    </layer-list>
-</item>
-    <item android:state_enabled="true">
-        <layer-list>
-            <item>
-                <vector android:height="24dp" android:viewportHeight="200"
-                    android:viewportWidth="200" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-                    <path android:fillColor="@color/fog" android:pathData="M69.28,82.82H79.86V66a20.14,20.14 0,1 1,40.28 0V82.82h10.58a27.1,27.1 0,0 1,6.45 0.78l2.7,0.66V66a39.87,39.87 0,0 0,-79.74 0V84.25l2.7,-0.66a27.61,27.61 0,0 1,6.45 -0.77Z"/>
-                    <path android:fillColor="@color/fog" android:pathData="M109.51,133.28v14.43a9.52,9.52 0,0 1,-19 0L90.51,133.28a16.45,16.45 0,1 1,19 0ZM130.72,86.74L69.28,86.74a25,25 0,0 0,-25 25L44.28,146.4a27.51,27.51 0,0 0,27.51 27.51h56.44a27.51,27.51 0,0 0,27.51 -27.51L155.74,111.75A25,25 0,0 0,130.72 86.74Z"/>
-                </vector>
-            </item>
-
-        </layer-list>
-    </item>
-</selector>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="4dp"
+    android:height="4dp"
+    android:viewportWidth="200"
+    android:viewportHeight="200">
+    <path
+        android:name="canvas_background"
+        android:pathData="M-1,-1 L581,-1 L581,401 L-1,401 L-1,-1 Z" />
+    <path
+        android:name="svg_1"
+        android:fillColor="@color/fog"
+        android:pathData="M68.7186,82.4914 L79.4884,82.4914 L79.4884,65.3696A20.501385,20.501385,0,1,1,120.491,65.3696 L120.491,82.4914 L131.261,82.4914A27.586272,27.586272,0,0,1,137.827,83.2854 L140.575,83.9572 L140.575,65.3695A40.585412,40.585412,0,0,0,59.4042,65.3695 L59.4042,83.947 L62.1526,83.2752A28.105423,28.105423,0,0,1,68.7183,82.4914 Z"
+        android:strokeWidth="1" />
+    <path
+        android:name="svg_2"
+        android:fillColor="@color/fog"
+        android:pathData="M109.67,133.857 L109.67,148.546 A9.690823,9.690823,0,0,1,90.3291,148.546L90.3291,133.857 A16.745173,16.745173,0,1,1,109.67,133.857 Z M131.261,86.4819L68.7185,86.4819 A25.448591,25.448591,0,0,0,43.2699,111.93 L43.2699,147.212A28.003629,28.003629,0,0,0,71.2735,175.216 L128.726,175.216A28.003629,28.003629,0,0,0,156.73,147.212 L156.73,111.94A25.448591,25.448591,0,0,0,131.261,86.4812 Z"
+        android:strokeWidth="1" />
+</vector>
