+++ app/src/main/res/drawable/ic_settings_reportissue.xml
@@ -1,75 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-<item android:state_pressed="true">
-    <layer-list>
-        <item>
-            <vector xmlns:android="http://schemas.android.com/apk/res/android"
-                android:width="34dp"
-                android:height="34dp"
-                android:viewportWidth="200"
-                android:viewportHeight="200">
-                <path
-                    android:fillColor="@color/fog"
-                    android:pathData="M 106.16,97.34 a 6.17,6.17 0 0 1 -12.33,0 V 72.71 a 6.17,6.17 0 0 1 12.33,0 z m
--6.16,30.8 a 6.16,6.16 0 1 1 6.17,-6.16 6.16,6.16 0 0 1 -6.17,6.16 z m 0,-80.07
-c -30.64,0 -55.45,22.43 -55.45,50.07 a 46.49,46.49 0 0 0 8.77,27 11.35,11.35 0 0
-1 0.88,1.33 6.68,6.68 0 0 1 0.89,3.28 c 0,0.45 -5.16,18.82 -5.16,18.82 a
-2.71,2.71 0 0 0 2.19,3.24 2.1,2.1 0 0 0 0.7,0.08 4.77,4.77 0 0 0 1.07,-0.16 l
-0.46,-0.2 14.65,-7.09 c 0.26,-0.13 2.6,-1.14 2.89,-1.27 a 9,9 0 0 1 3.23,-0.58
-9.17,9.17 0 0 1 3,0.48 c 0,0 0.06,0 0.06,0 0,0 0.28,0.1 0.43,0.16 6.67,2.56
-14,3.34 21.7,3.34 30.62,0 55.15,-20.79 55.15,-48.44 0,-27.65 -24.81,-50.06
--55.46,-50.06 z" />
-                </vector>
-        </item>
-
-    </layer-list>
-</item>
-<item android:state_hovered="true">
-    <layer-list>
-        <item>
-            <vector xmlns:android="http://schemas.android.com/apk/res/android"
-                android:width="34dp"
-                android:height="34dp"
-                android:viewportWidth="200"
-                android:viewportHeight="200">
-                <path
-                    android:fillColor="@color/fog"
-                    android:pathData="M 106.16,97.34 a 6.17,6.17 0 0 1 -12.33,0 V 72.71 a 6.17,6.17 0 0 1 12.33,0 z m
--6.16,30.8 a 6.16,6.16 0 1 1 6.17,-6.16 6.16,6.16 0 0 1 -6.17,6.16 z m 0,-80.07
-c -30.64,0 -55.45,22.43 -55.45,50.07 a 46.49,46.49 0 0 0 8.77,27 11.35,11.35 0 0
-1 0.88,1.33 6.68,6.68 0 0 1 0.89,3.28 c 0,0.45 -5.16,18.82 -5.16,18.82 a
-2.71,2.71 0 0 0 2.19,3.24 2.1,2.1 0 0 0 0.7,0.08 4.77,4.77 0 0 0 1.07,-0.16 l
-0.46,-0.2 14.65,-7.09 c 0.26,-0.13 2.6,-1.14 2.89,-1.27 a 9,9 0 0 1 3.23,-0.58
-9.17,9.17 0 0 1 3,0.48 c 0,0 0.06,0 0.06,0 0,0 0.28,0.1 0.43,0.16 6.67,2.56
-14,3.34 21.7,3.34 30.62,0 55.15,-20.79 55.15,-48.44 0,-27.65 -24.81,-50.06
--55.46,-50.06 z" />
-            </vector>
-        </item>
-
-    </layer-list>
-</item>
-<item android:state_enabled="true">
-    <layer-list>
-        <item>
-            <vector xmlns:android="http://schemas.android.com/apk/res/android"
-                android:width="34dp"
-                android:height="34dp"
-                android:viewportWidth="200"
-                android:viewportHeight="200">
-                <path
-                    android:fillColor="@color/fog"
-                    android:pathData="M 106.16,97.34 a 6.17,6.17 0 0 1 -12.33,0 V 72.71 a 6.17,6.17 0 0 1 12.33,0 z m
--6.16,30.8 a 6.16,6.16 0 1 1 6.17,-6.16 6.16,6.16 0 0 1 -6.17,6.16 z m 0,-80.07
-c -30.64,0 -55.45,22.43 -55.45,50.07 a 46.49,46.49 0 0 0 8.77,27 11.35,11.35 0 0
-1 0.88,1.33 6.68,6.68 0 0 1 0.89,3.28 c 0,0.45 -5.16,18.82 -5.16,18.82 a
-2.71,2.71 0 0 0 2.19,3.24 2.1,2.1 0 0 0 0.7,0.08 4.77,4.77 0 0 0 1.07,-0.16 l
-0.46,-0.2 14.65,-7.09 c 0.26,-0.13 2.6,-1.14 2.89,-1.27 a 9,9 0 0 1 3.23,-0.58
-9.17,9.17 0 0 1 3,0.48 c 0,0 0.06,0 0.06,0 0,0 0.28,0.1 0.43,0.16 6.67,2.56
-14,3.34 21.7,3.34 30.62,0 55.15,-20.79 55.15,-48.44 0,-27.65 -24.81,-50.06
--55.46,-50.06 z" />
-            </vector>
-        </item>
-
-    </layer-list>
-</item>
-</selector>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="200"
+    android:viewportHeight="200">
+    <path
+        android:name="canvas_background"
+        android:pathData="M-1,-1 L581,-1 L581,401 L-1,401 L-1,-1 Z" />
+    <path
+        android:name="svg_1"
+        android:fillColor="@color/fog"
+        android:pathData="M108.43,96.3819 A8.450858,8.450858,0,0,1,91.542,96.3819 L91.542,62.647A8.450858,8.450858,0,0,1,108.43,62.647 L108.43,96.3819 Z M99.9928,138.568A8.437161,8.437161,0,1,1,108.444,130.131A8.437161,8.437161,0,0,1,99.9931,138.568 Z M99.9928,28.899C58.0261,28.899,24.0447,59.6207,24.0447,97.4783A63.675912,63.675912,0,0,0,36.0567,134.459A15.545743,15.545743,0,0,1,37.262,136.281A9.149389,9.149389,0,0,1,38.481,140.774C38.481,141.39,31.4135,166.551,31.4135,166.551A3.711803,3.711803,0,0,0,34.4131,170.989A2.876305,2.876305,0,0,0,35.3719,171.099 A6.533321,6.533321,0,0,0,36.8374,170.88L37.4674,170.606 L57.5331,160.895C57.8892,160.717,61.0942,159.334,61.4914,159.156A12.327021,12.327021,0,0,1,65.9154,158.362A12.559865,12.559865,0,0,1,70.0244,159.019 S70.1066,159.019,70.1066,159.019S70.4901,159.156,70.6956,159.238 C79.8313,162.744,89.871,163.813,100.417,163.813C142.356,163.813,175.954,135.338,175.954,97.4662S141.973,28.9006,99.9922,28.9006 Z"
+        android:strokeWidth="1" />
+</vector>
