+++ app/src/main/res/layout/honeycomb_button.xml
@@ -1,33 +1,40 @@
 <?xml version="1.0" encoding="utf-8"?>
 <merge xmlns:android="http://schemas.android.com/apk/res/android">
-    <ImageView
-        android:id="@+id/settings_button_icon"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:src="@drawable/ic_settings_privacypolicy"
-        android:gravity="center" />
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical"
+        android:paddingTop="45dp">
+        <ImageView
+            android:id="@+id/settings_button_icon"
+            android:layout_width="24dp"
+            android:layout_height="24dp"
+            android:layout_gravity="center"
+            android:scaleType="fitXY"
+            android:src="@drawable/ic_settings_privacypolicy"
+            android:gravity="center" />
 
-    <TextView
-        android:id="@+id/settings_button_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_marginLeft="15dp"
-        android:layout_marginRight="15dp"
-        android:gravity="center"
-        android:text="@string/settings_privacy"
-        android:textSize="@dimen/text_smaller_size"
-        android:fontFamily="sans-serif-light"/>
-    <TextView
-        android:id="@+id/settings_secondary_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:layout_marginLeft="15dp"
-        android:layout_marginRight="15dp"
-        android:gravity="center"
-        android:text="Secondary text"
-        android:textSize="@dimen/text_smaller_size"
-        android:fontFamily="sans-serif-light"/>
+        <TextView
+            android:id="@+id/settings_button_text"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:layout_marginLeft="15dp"
+            android:layout_marginRight="15dp"
+            android:gravity="center"
+            android:text="@string/settings_privacy"
+            android:textSize="@dimen/text_smaller_size"
+            android:fontFamily="sans-serif-light"/>
+        <TextView
+            android:id="@+id/settings_secondary_text"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:layout_marginLeft="15dp"
+            android:layout_marginRight="15dp"
+            android:gravity="center"
+            android:text="Secondary text"
+            android:textSize="@dimen/text_smaller_size"
+            android:fontFamily="sans-serif-light"/>
+    </LinearLayout>
 </merge>
