+++ app/src/main/res/layout/honeycomb_switch.xml
@@ -7,7 +7,6 @@
         android:layout_gravity="center"
         android:fontFamily="sans-serif-light"
         android:gravity="center"
-        android:paddingTop="20dp"
         android:paddingBottom="5dp"
         android:textColor="@color/white"
         android:textSize="@dimen/text_smaller_size"
