+++ app/src/main/res/layout/options_language.xml
@@ -1,0 +1,70 @@
+<?xml version="1.0" encoding="utf-8"?>
+<merge xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <RelativeLayout
+        android:id="@+id/optionsLayout"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="@drawable/dialog_background"
+        android:paddingStart="30dp"
+        android:paddingTop="20dp"
+        android:paddingEnd="30dp"
+        android:paddingBottom="20dp">
+
+        <org.mozilla.vrbrowser.ui.views.UIButton
+            android:id="@+id/backButton"
+            style="?attr/navigationBarButtonStyle"
+            android:layout_width="40dp"
+            android:layout_height="40dp"
+            android:layout_alignParentStart="true"
+            android:layout_alignParentTop="true"
+            android:layout_centerVertical="true"
+            android:layout_marginBottom="10dp"
+            android:src="@drawable/ic_icon_back"
+            android:tint="@color/midnight" />
+
+        <TextView
+            android:id="@+id/voice_search_language_title"
+            style="@style/settingsDescription"
+            android:layout_below="@+id/backButton"
+            android:layout_marginBottom="10dp"
+            android:gravity="center_vertical"
+            android:text="@string/settings_language_choose_language_voice_search"
+            tools:text="@string/settings_language_choose_language_voice_search" />
+
+        <ScrollView
+            android:id="@+id/scrollbar"
+            style="@style/scrollbar"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:layout_above="@+id/resetButton"
+            android:layout_below="@+id/voice_search_language_title">
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginEnd="20dp"
+                android:orientation="vertical">
+
+                <org.mozilla.vrbrowser.ui.views.settings.RadioGroupVSetting
+                    android:id="@+id/languageRadio"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    app:layout="@layout/setting_radio_group_v"
+                    app:options="@array/developer_options_voice_search_languages"
+                    app:values="@array/developer_options_voice_search_languages_values" />
+
+            </LinearLayout>
+        </ScrollView>
+
+        <org.mozilla.vrbrowser.ui.views.settings.ButtonSetting
+            android:id="@+id/resetButton"
+            android:layout_width="match_parent"
+            android:layout_height="60dp"
+            android:layout_alignParentBottom="true"
+            app:buttonText="@string/developer_options_reset_button"
+            app:description="@string/language_options_reset" />
+    </RelativeLayout>
+</merge>
