+++ app/src/main/res/layout/setting_button.xml
@@ -6,11 +6,6 @@
         android:layout_height="60dp"
         android:orientation="vertical">
 
-        <View
-            android:layout_width="match_parent"
-            android:layout_height="0.5dp"
-            android:background="@color/iron" />
-
         <RelativeLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
