+++ app/src/main/res/values/attrs.xml
@@ -34,6 +34,7 @@
         <attr name="description" />
         <attr name="options" format="reference" />
         <attr name="values" format="reference" />
+        <attr name="layout" format="reference" />
     </declare-styleable>
 
     <declare-styleable name="EditSetting">
