+++ app/src/main/res/values/non_L10n.xml
@@ -21,6 +21,7 @@
     <string name="settings_key_env" translatable="false">settings_env</string>
     <string name="settings_key_pointer_color" translatable="false">settings_pointer_color</string>
     <string name="settings_key_msaa" translatable="false">settings_msaa</string>
+    <string name="settings_key_voice_search_language" translatable="false">settings_voice_search_language</string>
     <string name="private_browsing_support_url" translatable="false">https://support.mozilla.org/kb/private-mode-firefox-reality</string>
     <string name="settings_key_browser_world_width" translatable="false">settings_browser_world_width</string>
     <string name="settings_key_browser_world_height" translatable="false">settings_browser_world_height</string>
@@ -36,6 +37,17 @@
     <string name="crash_app_name" translatable="false">FirefoxReality</string>
     <string name="user_agent_override_file" translatable="false">userAgentOverride.json</string>
     <string name="about_bookmarks" translatable="false">about:bookmarks</string>
+    <string name="voice_app_id" translatable="false">fxr</string>
+
+    <string name="language_en_US" translatable="false">English (US)</string>
+    <string name="language_cmn_Hant_TW" translatable="false">繁體中文</string>
+    <string name="language_cmn_Hans_CN" translatable="false">简体中文</string>
+    <string name="language_ja_JP" translatable="false">日本語</string>
+    <string name="language_fr_FR" translatable="false">Français</string>
+    <string name="language_de_DE" translatable="false">Deutsch</string>
+    <string name="language_es_ES" translatable="false">Español</string>
+    <string name="language_ko_KR" translatable="false">한국어</string>
+    <string name="language_it_IT" translatable="false">Italiano</string>
 
     <!-- SEARCH ENGINES -->
     <string name="geolocation_api_url" translatable="false">https://location.services.mozilla.com/v1/country</string>
@@ -53,25 +65,4 @@
     <string name="search_yandex_by" translatable="false">https://yandex.by/search</string>
     <string name="search_yandex_tr" translatable="false">https://yandex.com.tr/search</string>
     <string name="search_yandex_kz" translatable="false">https://yandex.kz/search</string>
-
-    <!-- Settings -->
-    <string-array name="developer_options_environments" translatable="false">
-        <item>@string/developer_options_env_void</item>
-        <item>@string/developer_options_env_cave</item>
-        <item>@string/developer_options_env_meadow</item>
-    </string-array>
-    <string-array name="developer_options_pointer_color_options" translatable="false">
-        <item>@string/developer_options_pointer_white</item>
-        <item>@string/developer_options_pointer_purple</item>
-    </string-array>
-    <string-array name="developer_options_ua_options" translatable="false">
-        <item>@string/developer_options_ua_mobile</item>
-        <item>@string/developer_options_ua_desktop</item>
-        <item>@string/developer_options_ua_vr</item>
-    </string-array>
-    <string-array name="developer_options_msaa_options" translatable="false">
-        <item>@string/developer_options_msaa_disabled</item>
-        <item>@string/developer_options_msaa_2</item>
-        <item>@string/developer_options_msaa_4</item>
-    </string-array>
 </resources>
