--- app/src/main/res/values/settings_values.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string-array name="developer_options_environments_values">
-        <item>void</item>
-        <item>cave</item>
-        <item>meadow</item>
-    </string-array>
-    <integer-array name="developer_options_pointer_colors_values">
-        <item>0xFFFFFFFF</item>
-        <item>0xC27FFCFF</item>
-    </integer-array>
-    <integer-array name="developer_options_ua_mode_values">
-        <item>0</item>
-        <item>1</item>
-        <item>2</item>
-    </integer-array>
-</resources>
